+++ src/novo_backend/classes/RobotWrapper.py
@@ -5,7 +5,7 @@ from queue import Queue
 from modules.robot.main import robot_daemon
 
 class RobotWrapper:
-	def __init__(self, queue: Queue, timeout: int =0.25):
+	def __init__(self, queue: Queue, timeout: float = 0.25):
 		self.thread = threading.Thread(target=robot_daemon, args=(queue,))
 		self.timeout = timeout
